pyproject.toml
[project]
name = "orchestrator"
version = "0.1.0"
description = "Local LLM orchestrator with MCP tools, RAG, and telemetry ingestion"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.22.1",
    "fastapi>=0.129.0",
    "httpx>=0.28.0",
    "lancedb>=0.29.2",
    "mcp>=1.11.0",
    "ollama>=0.6.1",
    "pydantic>=2.0.0",
    "python-dotenv>=1.1.0",
    "transformers>=5.2.0",
    "uvicorn>=0.35.0",
]
